+++ packages/vanilla-components/src/core/helpers/isNumber.ts
@@ -1,0 +1,9 @@
+const isNumber = (value: unknown): boolean => {
+  if (value === null) {
+    return false
+  }
+
+  return ['number'].includes(typeof value)
+}
+
+export default isNumber